Version 1.5.24
Added Default Country setting to options
Backwards compatibility for older REAXML that had the img xml elements
contained an images element instead of the current that has them in the
objects element with floor plans and documents. Now if there are no img
xml elements inside the objects xml element, then it will look for them
in an images xml element.